+++ src/TraceabilitySystem.Application/Interfaces/IBaseService.cs
@@ -5,6 +5,6 @@ namespace TraceabilitySystem.Application.Interfaces;
 public interface IBaseService<TEntity, TDto> where TEntity : class
 {
     Task<PagedResult<TDto>> GetPagedAsync(int page, int pageSize, CancellationToken cancellationToken = default);
-    Task<TDto> GetByIdAsync(long id, CancellationToken cancellationToken = default);
-    Task DeleteAsync(long id, CancellationToken cancellationToken = default);
+    Task<TDto> GetByIdAsync(int id, CancellationToken cancellationToken = default);
+    Task DeleteAsync(int id, CancellationToken cancellationToken = default);
 }
